E2E kimi-linear

+++ kernels/pto/mega_kernel_kda.cpp
@@ -60,7 +60,8 @@ template <bool isAIVOnly = true>
 AICORE inline void SyncAllImpl()
 {
     pipe_barrier(PIPE_ALL);
-    if constexpr (isAIVOnly) {
+    if constexpr (isAIVOnly)
+    {
         ffts_cross_core_sync(PIPE_MTE3, GetffstMsg(0x0, SYNC_AIV_ONLY_ALL));
         wait_flag_dev(SYNC_AIV_ONLY_ALL);
         return;
@@ -85,29 +86,31 @@ AICORE void mega_permute_THK_to_HTK(
     __gm__ T *src, __gm__ T *dst, int64_t T_len, int32_t HV)
 {
 #if defined(__DAV_C220_VEC__)
-    if (get_subblockid() != 0) return;
+    if (get_subblockid() != 0)
+        return;
     set_mask_norm();
     set_vector_mask(-1, -1);
 
     auto cid = get_block_idx();
     auto block_num = get_block_num();
 
-    constexpr int32_t BLOCK = 128;          // tokens per UB tile
+    constexpr int32_t BLOCK = 128; // tokens per UB tile
     constexpr int32_t UB0 = 0;
 
     using UBTileDyn = Tile<TileType::Vec, T, BLOCK, KD, BLayout::RowMajor,
                            DYNAMIC, DYNAMIC, SLayout::NoneBox, 512, PadValue::Zero>;
-    using Gm2D   = Shape<1, 1, 1, DYNAMIC, DYNAMIC>;
-    using GmSrcS = Stride<1, 1, 1, DYNAMIC, 1>;   // row stride = HV*K (runtime; skip other heads)
-    using GmDstS = Stride<1, 1, 1, KD, 1>;        // contiguous
+    using Gm2D = Shape<1, 1, 1, DYNAMIC, DYNAMIC>;
+    using GmSrcS = Stride<1, 1, 1, DYNAMIC, 1>; // row stride = HV*K (runtime; skip other heads)
+    using GmDstS = Stride<1, 1, 1, KD, 1>;      // contiguous
     GmSrcS src_stride(HV * KD);
 
     int64_t num_tok_blocks = (T_len + BLOCK - 1) / BLOCK;
     int64_t total = static_cast<int64_t>(HV) * num_tok_blocks;
 
     for (int64_t wi = static_cast<int64_t>(cid); wi < total;
-         wi += static_cast<int64_t>(block_num)) {
-        int64_t h  = wi / num_tok_blocks;
+         wi += static_cast<int64_t>(block_num))
+    {
+        int64_t h = wi / num_tok_blocks;
         int64_t bi = wi % num_tok_blocks;
         int64_t t0 = bi * BLOCK;
         int32_t valid = (t0 + BLOCK <= T_len)
@@ -115,7 +118,9 @@ AICORE void mega_permute_THK_to_HTK(
                             : static_cast<int32_t>(T_len - t0);
 
         {
-            Gm2D gs; gs.shape[3] = valid; gs.shape[4] = KD;
+            Gm2D gs;
+            gs.shape[3] = valid;
+            gs.shape[4] = KD;
             GlobalTensor<T, Gm2D, GmSrcS> gm(
                 src + (t0 * static_cast<int64_t>(HV) + h) * KD, gs, src_stride);
             UBTileDyn ld(valid, KD);
@@ -125,7 +130,9 @@ AICORE void mega_permute_THK_to_HTK(
         set_flag(PIPE_MTE2, PIPE_MTE3, EVENT_ID0);
         wait_flag(PIPE_MTE2, PIPE_MTE3, EVENT_ID0);
         {
-            Gm2D gs; gs.shape[3] = valid; gs.shape[4] = KD;
+            Gm2D gs;
+            gs.shape[3] = valid;
+            gs.shape[4] = KD;
             GlobalTensor<T, Gm2D, GmDstS> gm(
                 dst + (h * T_len + t0) * KD, gs);
             UBTileDyn st(valid, KD);
@@ -147,35 +154,41 @@ AICORE void mega_permute_THK_to_HTK(
 // ===================================================================
 
 #define call_kernel _mk_unused_kda_gc
-namespace mk_gc {
+namespace mk_gc
+{
 #include "gate_cumsum_kda.cpp"
 }
 #undef call_kernel
 
 #define call_kernel _mk_unused_kda_kkt
-namespace mk_kkt {
+namespace mk_kkt
+{
 #include "kkt_kda.cpp"
 }
 #undef call_kernel
 
-namespace mk_solve {
+namespace mk_solve
+{
 #include "tri_inverse_impl.cpp"
 }
 
 #define call_kernel _mk_unused_kda_wy
-namespace mk_wy {
+namespace mk_wy
+{
 #include "wy_kda.cpp"
 }
 #undef call_kernel
 
 #define call_kernel _mk_unused_kda_h
-namespace mk_h {
+namespace mk_h
+{
 #include "chunk_h_kda.cpp"
 }
 #undef call_kernel
 
 #define call_kernel _mk_unused_kda_o
-namespace mk_o {
+namespace mk_o
+{
 #include "chunk_o_kda.cpp"
 }
 #undef call_kernel
@@ -206,34 +219,34 @@ AICORE void mega_solve_tril(
 // ===================================================================
 extern "C" __global__ AICORE void launch_mega_kernel_kda(
     // ── inputs ──────────────────────────────────────────────────────
-    __gm__ uint8_t *q_hm_ptr,      // [1, HV, T, K] fp16 (head-major, scaled)
-    __gm__ uint8_t *k_hm_ptr,      // [1, HV, T, K] fp16 (head-major)
-    __gm__ uint8_t *v_ptr,         // [1, T, HV, V] fp16 (BSND)
-    __gm__ uint8_t *g_in_ptr,      // [1, T, HV, K] fp16 (BSND, raw gate)
-    __gm__ uint8_t *beta_hm_ptr,   // [1, HV, T]    fp16 (head-major)
+    __gm__ uint8_t *q_hm_ptr,    // [1, HV, T, K] fp16 (head-major, scaled)
+    __gm__ uint8_t *k_hm_ptr,    // [1, HV, T, K] fp16 (head-major)
+    __gm__ uint8_t *v_ptr,       // [1, T, HV, V] fp16 (BSND)
+    __gm__ uint8_t *g_in_ptr,    // [1, T, HV, K] fp16 (BSND, raw gate)
+    __gm__ uint8_t *beta_hm_ptr, // [1, HV, T]    fp16 (head-major)
     // ── masks / constants ───────────────────────────────────────────
-    __gm__ uint8_t *mask_strict_ptr,  // [C, C] fp32 (rows >  cols)
-    __gm__ uint8_t *mask_incl_ptr,    // [C, C] fp32 (rows >= cols)
-    __gm__ uint8_t *minus_id_ptr,     // [C, C] fp16 (-I)
-    __gm__ uint8_t *cu_seqlens_ptr,   // int32
+    __gm__ uint8_t *mask_strict_ptr, // [C, C] fp32 (rows >  cols)
+    __gm__ uint8_t *mask_incl_ptr,   // [C, C] fp32 (rows >= cols)
+    __gm__ uint8_t *minus_id_ptr,    // [C, C] fp16 (-I)
+    __gm__ uint8_t *cu_seqlens_ptr,  // int32
     // ── output ──────────────────────────────────────────────────────
-    __gm__ uint8_t *o_ptr,         // [1, T, HV, V] fp16 (BSND)
+    __gm__ uint8_t *o_ptr, // [1, T, HV, V] fp16 (BSND)
     // ── intermediate buffers ────────────────────────────────────────
-    __gm__ uint8_t *g_sum_ptr,     // [1, T, HV, K] fp16 (BSND)
-    __gm__ uint8_t *g_cs_hm_ptr,   // [1, HV, T, K] fp16 (head-major)
-    __gm__ uint8_t *L_ptr,         // [1, T, HV, C] fp16
-    __gm__ uint8_t *A_inv_ptr,     // [1, T, HV, C] fp16
-    __gm__ uint8_t *u_ptr,         // [1, T, HV, V] fp16
-    __gm__ uint8_t *w_ptr,         // [1, T, HV, K] fp16
-    __gm__ uint8_t *s_ptr,         // [tc, HV, K, V] fp16
-    __gm__ uint8_t *v_corr_ptr,    // [1, T, HV, V] fp16
+    __gm__ uint8_t *g_sum_ptr,   // [1, T, HV, K] fp32 (BSND)
+    __gm__ uint8_t *g_cs_hm_ptr, // [1, HV, T, K] fp32 (head-major)
+    __gm__ uint8_t *L_ptr,       // [1, T, HV, C] fp16
+    __gm__ uint8_t *A_inv_ptr,   // [1, T, HV, C] fp16
+    __gm__ uint8_t *u_ptr,       // [1, T, HV, V] fp16
+    __gm__ uint8_t *w_ptr,       // [1, T, HV, K] fp16
+    __gm__ uint8_t *s_ptr,       // [tc, HV, K, V] fp16
+    __gm__ uint8_t *v_corr_ptr,  // [1, T, HV, V] fp16
     // ── per-core workspaces ─────────────────────────────────────────
-    __gm__ uint8_t *kkt_ws_in_ptr,   // [bd*2, 2C, K] fp16
-    __gm__ uint8_t *kkt_ws_out_ptr,  // [bd*2, C, C]  fp16
-    __gm__ uint8_t *wy_ws_a2_ptr,    // [bd, C, C]    fp16
-    __gm__ uint8_t *wy_ws_keff_ptr,  // [bd, C, K]    fp16
-    __gm__ uint8_t *h_ws_ptr,        // [bd*5, K, K]  fp16
-    __gm__ uint8_t *o_ws_ptr,        // [bd*7, K, K]  fp16
+    __gm__ uint8_t *kkt_ws_in_ptr,  // [bd*2, 2C, K] fp32 (stages exp(±g_cs))
+    __gm__ uint8_t *kkt_ws_out_ptr, // [bd*2, C, C]  fp32 (unmasked gated K·K^T)
+    __gm__ uint8_t *wy_ws_a2_ptr,   // [bd, C, C]    fp16
+    __gm__ uint8_t *wy_ws_keff_ptr, // [bd, C, K]    fp16
+    __gm__ uint8_t *h_ws_ptr,       // [bd*5, K, K]  fp16
+    __gm__ uint8_t *o_ws_ptr,       // [bd*7, K, K]  fp32 (gated q/k + GEMM I/O)
     // ── scalars ─────────────────────────────────────────────────────
     int64_t batch_size,
     int64_t seq_len,
@@ -248,12 +261,12 @@ extern "C" __global__ AICORE void launch_mega_kernel_kda(
     // compile-time, so the fused .so is head-count-agnostic like the staged ones.
     const int32_t HV = num_heads;
     constexpr int32_t KD = GDN_D;
-    constexpr int32_t C  = GDN_C;
+    constexpr int32_t C = GDN_C;
 
     // ── Stage 1: gate_cumsum (BSND -> BSND) ──────────────────────────
     mk_gc::gate_cumsum_kda_kernel<KD, C>(
         reinterpret_cast<__gm__ half *>(g_in_ptr),
-        reinterpret_cast<__gm__ half *>(g_sum_ptr),
+        reinterpret_cast<__gm__ float *>(g_sum_ptr),
         reinterpret_cast<__gm__ int32_t *>(cu_seqlens_ptr),
         batch_size, seq_len, HV, ffts_addr);
 
@@ -265,9 +278,9 @@ extern "C" __global__ AICORE void launch_mega_kernel_kda(
     SyncAllImpl<false>();
 
     // ── Stage 2: transpose g_sum -> head-major g_cs ──────────────────
-    mega_permute_THK_to_HTK<half, KD>(
-        reinterpret_cast<__gm__ half *>(g_sum_ptr),
-        reinterpret_cast<__gm__ half *>(g_cs_hm_ptr),
+    mega_permute_THK_to_HTK<float, KD>(
+        reinterpret_cast<__gm__ float *>(g_sum_ptr),
+        reinterpret_cast<__gm__ float *>(g_cs_hm_ptr),
         total_tokens, HV);
 
 #ifdef MEGA_STOP_AFTER_TRANSPOSE
@@ -280,11 +293,11 @@ extern "C" __global__ AICORE void launch_mega_kernel_kda(
     // ── Stage 3: kkt (gated K·K^T lower-tri matrix) ──────────────────
     mk_kkt::kkt_kda_kernel<KD, C>(
         reinterpret_cast<__gm__ half *>(k_hm_ptr),
-        reinterpret_cast<__gm__ half *>(g_cs_hm_ptr),
+        reinterpret_cast<__gm__ float *>(g_cs_hm_ptr),
         reinterpret_cast<__gm__ half *>(beta_hm_ptr),
         reinterpret_cast<__gm__ float *>(mask_strict_ptr),
-        reinterpret_cast<__gm__ half *>(kkt_ws_in_ptr),
-        reinterpret_cast<__gm__ half *>(kkt_ws_out_ptr),
+        reinterpret_cast<__gm__ float *>(kkt_ws_in_ptr),
+        reinterpret_cast<__gm__ float *>(kkt_ws_out_ptr),
         reinterpret_cast<__gm__ half *>(L_ptr),
         reinterpret_cast<__gm__ int32_t *>(cu_seqlens_ptr),
         batch_size, seq_len, total_tokens, HV, ffts_addr);
@@ -316,7 +329,7 @@ extern "C" __global__ AICORE void launch_mega_kernel_kda(
         reinterpret_cast<__gm__ half *>(k_hm_ptr),
         reinterpret_cast<__gm__ half *>(v_ptr),
         reinterpret_cast<__gm__ half *>(beta_hm_ptr),
-        reinterpret_cast<__gm__ half *>(g_cs_hm_ptr),
+        reinterpret_cast<__gm__ float *>(g_cs_hm_ptr),
         reinterpret_cast<__gm__ half *>(A_inv_ptr),
         reinterpret_cast<__gm__ half *>(wy_ws_a2_ptr),
         reinterpret_cast<__gm__ half *>(wy_ws_keff_ptr),
@@ -337,7 +350,7 @@ extern "C" __global__ AICORE void launch_mega_kernel_kda(
         reinterpret_cast<__gm__ half *>(k_hm_ptr),
         reinterpret_cast<__gm__ half *>(w_ptr),
         reinterpret_cast<__gm__ half *>(u_ptr),
-        reinterpret_cast<__gm__ half *>(g_cs_hm_ptr),
+        reinterpret_cast<__gm__ float *>(g_cs_hm_ptr),
         reinterpret_cast<__gm__ half *>(s_ptr),
         reinterpret_cast<__gm__ half *>(v_corr_ptr),
         reinterpret_cast<__gm__ half *>(h_ws_ptr),
@@ -357,9 +370,9 @@ extern "C" __global__ AICORE void launch_mega_kernel_kda(
         reinterpret_cast<__gm__ half *>(k_hm_ptr),
         reinterpret_cast<__gm__ half *>(v_corr_ptr),
         reinterpret_cast<__gm__ half *>(s_ptr),
-        reinterpret_cast<__gm__ half *>(g_cs_hm_ptr),
+        reinterpret_cast<__gm__ float *>(g_cs_hm_ptr),
         reinterpret_cast<__gm__ float *>(mask_incl_ptr),
-        reinterpret_cast<__gm__ half *>(o_ws_ptr),
+        reinterpret_cast<__gm__ float *>(o_ws_ptr),
         reinterpret_cast<__gm__ half *>(o_ptr),
         reinterpret_cast<__gm__ int32_t *>(cu_seqlens_ptr),
         batch_size, seq_len, total_tokens, HV, ffts_addr);
